+++ src/types/weighted.rs
@@ -3,5 +3,5 @@ use serde::{Deserialize, Serialize};
 #[derive(Serialize, Deserialize, Debug)]
 pub struct LexiconWeighted {
     pub term: String,
-    pub weight: f32,
+    pub weight: f64,
 }
